pyproject.toml
[tool.poetry]
name = "module-virtual-machine"
version = "0.1.0"
description = ""
authors = ["Lucas <lhenriquerod@dxc.com>"]
readme = "README.md"
packages = [{include = "module_virtual_machine"}]

[tool.poetry.dependencies]
python = "^3.10"
dynaconf = "^3.1.12"
azure-mgmt-compute = "^30.0.0"
azure-identity = "^1.13.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pylint = "^2.17.4"
pytest-cov = "^4.1.0"
bandit = "^1.7.5"
safety = "^2.3.5"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
